# specs/001-babble-web-app/contracts/api.yaml
openapi: 3.1.0
info:
  title: Prompt Babbler API
  description: |
    Backend API for Prompt Babbler. V1 responsibilities:
    1. Proxy LLM calls to Azure OpenAI (CORS workaround)
    2. Proxy STT calls to Azure OpenAI (CORS workaround)
    3. Manage LLM/STT settings (persist to local config file)
    4. Health checks (via Aspire ServiceDefaults)
  version: 1.0.0
  contact:
    name: Prompt Babbler
    url: https://github.com/PlagueHO/prompt-babbler
servers:
  - url: http://localhost:5000
    description: Local development (Aspire-assigned, binds to localhost/127.0.0.1 only)
paths:
...
  /api/settings:
    get:
      operationId: getSettings
      summary: Get current LLM settings
      description: |
        Returns the current LLM configuration. The API key is masked —
        only the last 4 characters are returned as `apiKeyHint`.
      tags:
        - Settings
      responses:
        '200':
          description: Current LLM settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmSettingsResponse'
              examples:
                configured:
                  summary: Settings are configured
                  value:
                    endpoint: "https://my-resource.openai.azure.com/"
                    apiKeyHint: "...a1b2"
                    deploymentName: "gpt-4o-mini"
                    sttDeploymentName: "gpt-4o-transcribe"
                    isConfigured: true
                notConfigured:
                  summary: Settings not yet configured
                  value:
                    endpoint: ""
                    apiKeyHint: ""
                    deploymentName: ""
                    sttDeploymentName: ""
                    isConfigured: false

    put:
      operationId: updateSettings
      summary: Save LLM settings
      description: |
        Saves the LLM configuration to the backend config file
        (`~/.prompt-babbler/settings.json`). Validates the endpoint format
        and that all required fields are present.
      tags:
        - Settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmSettingsSaveRequest'
            example:
              endpoint: "https://my-resource.openai.azure.com/"
              apiKey: "sk-abc123def456..."
              deploymentName: "gpt-4o-mini"
              sttDeploymentName: "gpt-4o-transcribe"
      responses:
        '200':
          description: Settings saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmSettingsResponse'
        '400':
          description: Invalid settings (bad URL format, missing fields)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/settings/test:
    post:
      operationId: testConnection
      summary: Test LLM connection
      description: |
        Sends a lightweight request to the configured Azure OpenAI endpoint
        to verify that the endpoint is reachable and the credentials are valid.
        Uses the currently saved settings (not request body).
      tags:
        - Settings
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestConnectionResponse'
              examples:
                success:
                  summary: Connection successful
                  value:
                    success: true
                    message: "Successfully connected to Azure OpenAI endpoint."
                    latencyMs: 234
                failure:
                  summary: Connection failed
                  value:
                    success: false
                    message: "Authentication failed. Please verify your API key."
                    latencyMs: null
        '422':
          description: Settings not configured
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    LlmSettingsResponse:
      type: object
      required:
        - endpoint
        - apiKeyHint
        - deploymentName
        - sttDeploymentName
        - isConfigured
      properties:
        endpoint:
          type: string
          format: uri
          description: Azure OpenAI endpoint URL.
        apiKeyHint:
          type: string
          description: Masked API key showing only the last 4 characters (e.g., "...a1b2"). Empty if not configured.
        deploymentName:
          type: string
          description: Model deployment name for LLM/prompt generation.
        sttDeploymentName:
          type: string
          description: Model deployment name for speech-to-text (e.g., gpt-4o-transcribe).
        isConfigured:
          type: boolean
          description: Whether all required settings have been provided.
    LlmSettingsSaveRequest:
      type: object
      required:
        - endpoint
        - apiKey
        - deploymentName
        - sttDeploymentName
      properties:
        endpoint:
          type: string
          format: uri
          description: Azure OpenAI or Azure AI Foundry endpoint URL (must be HTTPS).
        apiKey:
          type: string
          minLength: 1
          description: API key for Azure OpenAI authentication.
        deploymentName:
          type: string
          minLength: 1
          maxLength: 64
          description: The deployment name of the LLM model to use for prompt generation.
        sttDeploymentName:
          type: string
          minLength: 1
          maxLength: 64
          description: The deployment name of the STT model to use for speech-to-text (e.g., gpt-4o-transcribe).
    TestConnectionResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: Whether the connection test succeeded.
        message:
          type: string
          description: Human-readable result message.
        latencyMs:
          type: integer
          nullable: true
          description: Round-trip latency in milliseconds (null if connection failed).
    ProblemDetails:
      type: object
      description: RFC 9457 Problem Details response.
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type.
        title:
          type: string
          description: Short human-readable summary.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI identifying the specific occurrence.
...